requirements.txt

Flask
gunicorn
python-dotenv
pymongo
pandas
openpyxl
python-decouple
